requirements.txt
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
yfinance>=0.2.28
pandas-ta>=0.3.14b
ta>=0.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
python-dotenv>=1.0.0